src/python/django-access-dashboard/requirements.txt
Django==1.7
jsonfield==1.0.0
django-cors-headers
djangorestframework>=2.4.3
django-filter
dj-config-url
markdown
clitool
python-dateutil>2.0
pytz
#python-dateutil==1.5  # for Python 2.x series

# Choose one of them for your production environment.
#mysql-connector-python
#PyMySQL
#py-postgresql

# Activate this if you use AWS.
#awscli

